refactor(global): move interceptor and resolver from user module to global module
User 모듈에 있던 공통 interceptor, resolver를 Global 모듈로 이동
모듈 간 중복 제거 및 재사용성 향상 목적
