+++ account/src/main/java/account/domain/account/controller/model/TokenValidationResponse.java
@@ -1,6 +1,6 @@
 package account.domain.account.controller.model;
 
-import com.example.user.domain.user.repository.enums.UserRole;
+import com.example.global.resolver.UserRole;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
